Drop redundant samples
